+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 This is the changelog for [AWS Architect](readme.md).
 
+## 4.2 ##
+* Include cloud formation deployment for standard resources.
+* Added example cloud formation template to template service.
+
 ## 4.1 ##
 * Add `+AwsArchitect.prototype.PublishLambdaArtifactPromise` to perform the action to deploy a microservice zip package to S3.
 * Allow specifying S3 artifacts cache-control times as overrides.